# feat(frontend): improved transactions list query performance